Squelch some compiler warnings.
Cast an ssize_t value to unsigned int before subtracting it from an
unsigned int.  The value is already known to be >= 0 and <= the value
from which we're subtracting it, and that latter value is an unsigned
int, so it's guaranteed to fit into an unsigned int.  (ssize_t can be
bigger than unsigned int.)

Reviewed-on: https://code.wireshark.org/review/21114
(cherry picked from commit 237d23dc7355f066d772ba6ad184d5f99c25fa5a)
Reviewed-on: https://code.wireshark.org/review/21115

+++ rawshark.c
@@ -846,7 +846,7 @@ main(int argc, char *argv[])
                     cmdarg_err("Not enough bytes for pcap header.");
                     exit(2);
                 }
-                bytes_left -= bytes;
+                bytes_left -= (unsigned int)bytes;
             }
         }
 
@@ -926,7 +926,7 @@ raw_pipe_read(struct wtap_pkthdr *phdr, guchar * pd, int *err, gchar **err_info,
             *err_info = NULL;
             return FALSE;
         }
-        bytes_needed -= bytes_read;
+        bytes_needed -= (unsigned int)bytes_read;
         *data_offset += bytes_read;
         ptr += bytes_read;
     }
@@ -972,7 +972,7 @@ raw_pipe_read(struct wtap_pkthdr *phdr, guchar * pd, int *err, gchar **err_info,
             *err_info = NULL;
             return FALSE;
         }
-        bytes_needed -= bytes_read;
+        bytes_needed -= (unsigned int)bytes_read;
         *data_offset += bytes_read;
         ptr += bytes_read;
     }
